some changes, i think i startint realize that
