+++ arakelyan.daniil/S1/main.cpp
@@ -3,10 +3,8 @@
 #include "binList.hpp"
 #include "binList.cpp"
 
-#include "iterator.hpp"
-#include "iterator.cpp"
-
 #include "node.hpp"
+#include "node.cpp"
 
 int main()
 {
@@ -30,7 +28,7 @@ int main()
   list.push_back(16);
   list.printList(std::cout);
   std::cout << "\n";
-
+  
   list.pop_back();
   list.printList(std::cout);
   std::cout << "\n";
